backend/requirements.txt
# Phronesis LEX Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Rate Limiting
slowapi>=0.1.9

# Database
aiosqlite>=0.19.0
asyncpg>=0.29.0

# AI
anthropic>=0.18.0

# Document Processing
PyMuPDF>=1.24.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0

# FCIP Analysis Engine Dependencies
rapidfuzz>=3.6.0
parsedatetime>=2.6
python-dateutil>=2.8.2
scipy>=1.12.0
numpy>=1.26.0
